@board parameter added to sql procedures Ceiling for Price field added on SyncSejourRates export

+++ MaratukAdmin/Repositories/Concrete/Sansejour/ContractExportRepository.cs
@@ -18,6 +18,7 @@ using Microsoft.EntityFrameworkCore.Query;
 using MaratukAdmin.Dto.Response.Sansejour;
 using MaratukAdmin.Entities;
 using MaratukAdmin.Managers.Abstract.Sansejour;
+using MaratukAdmin.Repositories.Abstract;
 
 namespace MaratukAdmin.Repositories.Concrete.Sansejour
 {
@@ -98,7 +99,7 @@ namespace MaratukAdmin.Repositories.Concrete.Sansejour
         public async Task<bool> DeleteSyncSejourContractsByDateAsync(DateTime exportDate)
         {
             var entity = await GetSyncSejourContractsByDateAsync(exportDate);
-            if (entity != null)
+            if (entity != null && entity.Count > 0)
             {
                 _dbContext.RemoveRange(entity);
                 await _dbContext.SaveChangesAsync();
@@ -121,6 +122,139 @@ namespace MaratukAdmin.Repositories.Concrete.Sansejour
         }
 
 
+        #region *** HotelBoard ***
+        //public async Task<List<HotelBoard>> GetHotelBoardsByCodeAsync(string? code = null)
+        //{
+        //    try
+        //    {
+        //        return await _dbContext.HotelBoard.Where(c => c.Board == ((code == null) ? c.Board : code)).ToListAsync();
+        //    }
+        //    catch (Exception)
+        //    {
+        //        throw;
+        //    }
+        //}
+        //public async Task<List<HotelBoard>> GetHotelBoardsFromRatesBySyncDateAsync(DateTime syncDate)
+        //{
+        //    try
+        //    {
+        //        return await _dbContext.SyncSejourRate
+        //                    .Where(r => r.SyncDate == syncDate)
+        //                    .OrderBy(r => r.Board)
+        //                    .Select(r => new HotelBoard()
+        //                    {
+        //                        Board= r.Board ?? "",
+        //                        BoardDesc = r.BoardDesc ?? ""
+        //                    })
+        //                    .Distinct()
+        //                    .ToListAsync();
+        //    }
+        //    catch (Exception)
+        //    {
+        //        throw;
+        //    }
+        //}
+
+        //public async Task AddNewHotelBoardsAsync(List<HotelBoard> hotelBoards)
+        //{
+        //    try
+        //    {
+        //        var existingRecords = await GetHashSyncSejourAccomodationTypeByCodeAsync();
+
+        //        foreach (var accmdRow in hotelBoards)
+        //        {
+        //            string accmdMenTypeCode = accmdRow.Code;
+        //            string accmdMenTypeName = accmdRow.Name;
+
+        //            string combinedKey = $"{accmdMenTypeCode}-{accmdMenTypeName}";
+
+        //            // Check if record is unique
+        //            if (!existingRecords.Contains(combinedKey))
+        //            {
+        //                SyncSejourAccomodationType newRecord = new()
+        //                {
+        //                    Code = accmdMenTypeCode,
+        //                    Name = accmdMenTypeName,
+        //                    Description = "",
+        //                    IsDeleted = false,
+        //                    IsDescribed = false,
+        //                    UpdateDate = DateTime.Now
+        //                };
+
+        //                await _dbContext.SyncSejourAccomodationType.AddAsync(newRecord);
+        //                existingRecords.Add(combinedKey);
+        //            }
+        //        }
+
+        //        await _dbContext.SaveChangesAsync();
+        //    }
+        //    catch (Exception)
+        //    {
+        //        throw;
+        //    }
+        //}
+
+        //public async Task<List<SyncSejourHotel>> GetHotelBoardsAsync(DateTime syncDate, string? hotelCode = null)
+        //{
+        //    try
+        //    {
+        //        return await _dbContext.SyncSejourHotel.Where(c => c.SyncDate == syncDate && c.HotelCode == (hotelCode ?? c.HotelCode)).ToListAsync();
+        //    }
+        //    catch (Exception)
+        //    {
+        //        throw;
+        //    }
+        //}
+        //public async Task<bool> DeleteHotelBoardByCodeAsync(DateTime exportDate, string? hotelCode = null)
+        //{
+        //    var entity = await GetHotelBoardsAsync(exportDate, hotelCode);
+        //    if (entity != null && entity.Count > 0)
+        //    {
+        //        _dbContext.RemoveRange(entity);
+        //        await _dbContext.SaveChangesAsync();
+        //        return true;
+        //    }
+
+        //    return false;
+        //}
+
+        //public async Task<bool> DeleteSyncSejourHotelsByDateRAWAsync(DateTime exportDate, string? hotelCode = null)
+        //{
+        //    try
+        //    {
+        //        string sqlQuery = @$"DELETE FROM {nameof(SyncSejourHotel)} 
+        //                            WHERE {nameof(SyncSejourHotel.SyncDate)} = '{exportDate.ToString("yyyy-MM-dd")}'
+        //                            AND {nameof(SyncSejourHotel.HotelCode)} = " + ((hotelCode == null)
+        //                                                                        ? nameof(SyncSejourHotel.HotelCode)
+        //                                                                        : "'" + hotelCode + "'");
+        //        await _dbContext.Database.ExecuteSqlRawAsync(sqlQuery);
+
+        //        await _dbContext.SaveChangesAsync();
+        //        return true;
+        //    }
+        //    catch (Exception)
+        //    {
+        //        return false;
+        //    }
+        //}
+
+        //public async Task AddNewSejourHotelsAsync(SyncSejourHotel syncHotel)
+        //{
+        //    try
+        //    {
+        //        var hotel = _dbContext.SyncSejourHotel.Where(c => c.HotelCode == syncHotel.HotelCode).FirstOrDefault();
+        //        if (hotel == null)
+        //        {
+        //            await _dbContext.SyncSejourHotel.AddAsync(syncHotel);
+        //            await _dbContext.SaveChangesAsync();
+        //        }
+        //    }
+        //    catch (Exception)
+        //    {
+        //        throw;
+        //    }
+        //}
+        #endregion
 
         #region *** SyncSejourHotel ***
         public async Task<List<SyncSejourHotel>> GetSyncSejourHotelsByDateAsync(DateTime syncDate, string? hotelCode = null)
@@ -137,7 +271,7 @@ namespace MaratukAdmin.Repositories.Concrete.Sansejour
         public async Task<bool> DeleteSyncSejourHotelsByDateAsync(DateTime exportDate, string? hotelCode = null)
         {
             var entity = await GetSyncSejourHotelsByDateAsync(exportDate, hotelCode);
-            if (entity != null)
+            if (entity != null && entity.Count > 0)
             {
                 _dbContext.RemoveRange(entity);
                 await _dbContext.SaveChangesAsync();
@@ -152,7 +286,7 @@ namespace MaratukAdmin.Repositories.Concrete.Sansejour
             try
             {
                 string sqlQuery = @$"DELETE FROM {nameof(SyncSejourHotel)} 
-                                    WHERE {nameof(SyncSejourHotel.SyncDate)} = '{exportDate}'
+                                    WHERE {nameof(SyncSejourHotel.SyncDate)} = '{exportDate.ToString("yyyy-MM-dd")}'
                                     AND {nameof(SyncSejourHotel.HotelCode)} = " + ((hotelCode == null)
                                                                                 ? nameof(SyncSejourHotel.HotelCode)
                                                                                 : "'" + hotelCode + "'");
@@ -200,7 +334,7 @@ namespace MaratukAdmin.Repositories.Concrete.Sansejour
         public async Task<bool> DeleteSyncSejourSpoAppOrdersByDateAsync(DateTime exportDate, string? hotelCode = null)
         {
             var entity = await GetSyncSejourSpoAppOrdersByDateAsync(exportDate, hotelCode);
-            if (entity != null)
+            if (entity != null && entity.Count > 0)
             {
                 _dbContext.RemoveRange(entity);
                 await _dbContext.SaveChangesAsync();
@@ -213,7 +347,7 @@ namespace MaratukAdmin.Repositories.Concrete.Sansejour
             try
             {
                 string sqlQuery = @$"DELETE FROM {nameof(SyncSejourSpoAppOrder)} 
-                                    WHERE {nameof(SyncSejourSpoAppOrder.SyncDate)} = '{exportDate}'
+                                    WHERE {nameof(SyncSejourSpoAppOrder.SyncDate)} = '{exportDate.ToString("yyyy-MM-dd")}'
                                     AND {nameof(SyncSejourSpoAppOrder.HotelCode)} = " + ((hotelCode == null)
                                                                                 ? nameof(SyncSejourSpoAppOrder.HotelCode)
                                                                                 : "'" + hotelCode + "'");
@@ -256,7 +390,7 @@ namespace MaratukAdmin.Repositories.Concrete.Sansejour
         public async Task<bool> DeleteSyncSejourSpecialOffersByDateAsync(DateTime exportDate, string? hotelCode = null)
         {
             var entity = await GetSyncSejourSpecialOffersByDateAsync(exportDate, hotelCode);
-            if (entity != null)
+            if (entity != null && entity.Count > 0)
             {
                 _dbContext.RemoveRange(entity);
                 await _dbContext.SaveChangesAsync();
@@ -270,7 +404,7 @@ namespace MaratukAdmin.Repositories.Concrete.Sansejour
             try
             {
                 string sqlQuery = @$"DELETE FROM {nameof(SyncSejourSpecialOffer)} 
-                                    WHERE {nameof(SyncSejourSpecialOffer.SyncDate)} = '{exportDate}'
+                                    WHERE {nameof(SyncSejourSpecialOffer.SyncDate)} = '{exportDate.ToString("yyyy-MM-dd")}'
                                     AND {nameof(SyncSejourSpecialOffer.HotelCode)} = " + ((hotelCode == null)
                                                                                 ? nameof(SyncSejourSpecialOffer.HotelCode)
                                                                                 : "'" + hotelCode + "'");
@@ -313,7 +447,7 @@ namespace MaratukAdmin.Repositories.Concrete.Sansejour
         public async Task<bool> DeleteSyncSejourRatesByDateAsync(DateTime exportDate, string? hotelCode = null)
         {
             var entity = await GetSyncSejourRatesByDateAsync(exportDate, hotelCode);
-            if (entity != null)
+            if (entity != null && entity.Count > 0)
             {
                 _dbContext.RemoveRange(entity);
                 await _dbContext.SaveChangesAsync();
@@ -326,7 +460,7 @@ namespace MaratukAdmin.Repositories.Concrete.Sansejour
             try
             {
                 string sqlQuery = @$"DELETE FROM {nameof(SyncSejourRate)} 
-                                    WHERE {nameof(SyncSejourRate.SyncDate)} = '{exportDate}'
+                                    WHERE {nameof(SyncSejourRate.SyncDate)} = '{exportDate.ToString("yyyy-MM-dd")}'
                                     AND {nameof(SyncSejourRate.HotelCode)} = " + ((hotelCode == null)
                                                                                 ? nameof(SyncSejourRate.HotelCode)
                                                                                 : "'" + hotelCode + "'");
@@ -473,7 +607,7 @@ namespace MaratukAdmin.Repositories.Concrete.Sansejour
             try
             {
                 //string sqlQuery = @$"DELETE FROM {nameof(SyncSejourAccomodationTypes)} 
-                //                    WHERE {nameof(SyncSejourAccomodationTypes.SyncDate)} = '{exportDate}'
+                //                    WHERE {nameof(SyncSejourAccomodationTypes.SyncDate)} = '{exportDate.ToString("yyyy-MM-dd")}'
                 //                    AND {nameof(SyncSejourAccomodationTypes.HotelCode)} = " + ((hotelCode == null)
                 //                                                                ? nameof(SyncSejourAccomodationTypes.HotelCode)
                 //                                                                : "'" + hotelCode + "'");
@@ -641,9 +775,59 @@ namespace MaratukAdmin.Repositories.Concrete.Sansejour
         #region *** SEARCH ***
         public async Task<DateTime?> GetMaxSyncDateAsync()
         {
+            DateTime? maxDate = DateTime.MinValue;
             try
             {
-                return await Task.FromResult(_dbContext.SyncSejourContractExportView.Max(e => e.ExportDate));
+                //return await Task.FromResult(_dbContext.SyncSejourContractExportView.Max(e => e.ExportDate));
+                maxDate = _dbContext.SyncSejourContractExportView.Max(e => e.ExportDate);
+            }
+            catch (Exception)
+            {
+                throw;
+            }
+
+            return await Task.FromResult(maxDate);
+        }
+        public async Task<DateTime?> GetMaxSyncDateAsyncNew()
+        {
+            DateTime? maxDate = DateTime.MinValue;
+            try
+            {
+                var strategy = _dbContext.Database.CreateExecutionStrategy();
+
+                await strategy.ExecuteAsync(async () =>
+                {
+                    using (var transaction = await _dbContext.Database.BeginTransactionAsync())
+                    {
+                        try
+                        {
+                            //return await Task.FromResult(_dbContext.SyncSejourContractExportView.Max(e => e.ExportDate));
+                            maxDate = _dbContext.SyncSejourContractExportView.Max(e => e.ExportDate);
+
+                            transaction.Commit();
+                        }
+                        catch (Exception)
+                        {
+                            transaction.Rollback();
+                            throw;
+                        }
+                    }
+                });
+            }
+            catch (Exception)
+            {
+                throw;
+            }
+
+            return await Task.FromResult(maxDate);
+        }
+
+        public async Task<DateTime?> GetMaxSyncDateAsyncNew(MaratukDbContext dbContext)
+        {
+            try
+            {
+                //return await Task.FromResult(_dbContext.SyncSejourContractExportView.Max(e => e.ExportDate));
+                return await Task.FromResult(dbContext.SyncSejourContractExportView.Max(e => e.ExportDate));
             }
             catch (Exception)
             {
@@ -738,6 +922,7 @@ namespace MaratukAdmin.Repositories.Concrete.Sansejour
                 DateTime? exportDate = (searchRequest.ExportDate == null) ? await GetMaxSyncDateAsync() : searchRequest.ExportDate;
                 DateTime? accomodationDateFrom = searchRequest.AccomodationDateFrom;
                 DateTime? accomodationDateTo = searchRequest.AccomodationDateTo;
+                string? board = searchRequest.Board;
                 int roomPax = searchRequest.TotalCount;
                 int adultPax = searchRequest.AdultCount;
                 int? childPax = searchRequest.ChildCount;
@@ -760,13 +945,14 @@ namespace MaratukAdmin.Repositories.Concrete.Sansejour
 
                 //var result = await _dbContext.SyncSejourRate.FromSqlRaw("EXEC dbo.Sp_Search_Room " +
                 var result = await _dbContext.RoomSearchResponse.FromSqlRaw("EXEC dbo.Sp_Search_Room " +
-                                                                    "@exportDate, @accomodationDateFrom, @accomodationDateTo, " +
+                                                                    "@exportDate, @accomodationDateFrom, @accomodationDateTo, @board, " +
                                                                     "@roomPax, @adultPax, @childPax, " +
                                                                     "@childAge1, @childAge2, @childAge3, @childAge4, @childAge5, " +
                                                                     "@hotelCodes, @pageNumber, @pageSize",
                                                                     new SqlParameter("exportDate", exportDate),
                                                                     new SqlParameter("accomodationDateFrom", accomodationDateFrom),
                                                                     new SqlParameter("accomodationDateTo", accomodationDateTo),
+                                                                    new SqlParameter("board", board),
                                                                     new SqlParameter("roomPax", roomPax),
                                                                     new SqlParameter("adultPax", adultPax),
                                                                     new SqlParameter("childPax", childPax),
@@ -1051,6 +1237,7 @@ namespace MaratukAdmin.Repositories.Concrete.Sansejour
                 DateTime? exportDate = (searchRequest.ExportDate == null) ? await GetMaxSyncDateAsync() : searchRequest.ExportDate;
                 DateTime? accomodationDateFrom = searchRequest.AccomodationDateFrom;
                 DateTime? accomodationDateTo = searchRequest.AccomodationDateTo;
+                string? board = searchRequest.Board;
                 int roomPax = searchRequest.TotalCount;
                 int adultPax = searchRequest.AdultCount;
                 int? childPax = searchRequest.ChildCount;
@@ -1065,13 +1252,14 @@ namespace MaratukAdmin.Repositories.Concrete.Sansejour
 
                 //var result = await _dbContext.SyncSejourRate.FromSqlRaw("EXEC dbo.Sp_Search_Room_LowestPrices " +
                 var result = await _dbContext.RoomSearchResponseLowestPrices.FromSqlRaw("EXEC dbo.Sp_Search_Room_LowestPrices " +
-                                                                        "@exportDate, @accomodationDateFrom, @accomodationDateTo, " +
+                                                                        "@exportDate, @accomodationDateFrom, @accomodationDateTo, @board, " +
                                                                         "@roomPax, @adultPax, @childPax, " +
                                                                         "@childAge1, @childAge2, @childAge3, @childAge4, @childAge5, " +
                                                                         "@hotelCodes, @pageNumber, @pageSize",
                                                                         new SqlParameter("exportDate", exportDate),
                                                                         new SqlParameter("accomodationDateFrom", accomodationDateFrom),
                                                                         new SqlParameter("accomodationDateTo", accomodationDateTo),
+                                                                        new SqlParameter("@board", board),
                                                                         new SqlParameter("roomPax", roomPax),
                                                                         new SqlParameter("adultPax", adultPax),
                                                                         new SqlParameter("childPax", childPax),
